Sync backend manually w/ branch "backend"

+++ backend/channel/Communication.go
@@ -5,7 +5,7 @@ type Communication struct {
 	Value	string	`json:"value"`
 }
 
-func (s *Server) GetCommunication(commID int) (com []Communication, err error) {
+func (s *Server) GetCommunication(commID int64) (com []Communication, err error) {
 	com = make([]Communication, 0)
 	get, err := s.dataBase.Prepare("SELECT C.val, CT.val FROM COMM AS C ,COMM_TYPE AS CT WHERE C.user_id = ? AND C.type_id = CT.type_id")
 	if err != nil {
